+++ packages/map-interface/CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file. The format
 is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this
 project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.0] - 2026-01-26
+
+- Update peer dependencies to React 18
+- Convert last remaining Stylus to Sass
+- Update bundling process to `@macrostrat/web-components-bundler`
+- Make styles optional: import of `@macrostrat/map-interface/styles.css` or
+  `@macrostrat/map-interface/dist/map-interface.css` is now required for styles
+
 ## [1.6.0] - 2025-11-28
 
 - Update XDD panel components to remove Typescript